Segunda entrega

+++ app/src/main/java/com/manoelfreitas/imdmarket/ui/screen/main/product/deleteproduct/DeleteProductScreen.kt
@@ -16,7 +16,7 @@ import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.sp
-import com.manoelfreitas.imdmarket.ui.screen.main.product.ProductViewModel
+import com.manoelfreitas.imdmarket.product.viewModel.ProductViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -90,7 +90,7 @@ fun DeleteProductScreen(navController: NavController){
 
 
 fun deleteProduct(context: Context, productCode: Int?) {
-    if(productCode != null) {
+    if(productCode != null && ProductViewModel().isProductExists(productCode)) {
         ProductViewModel().deleteProduct(productCode)
 
         Toast.makeText(
@@ -101,7 +101,7 @@ fun deleteProduct(context: Context, productCode: Int?) {
     }else {
         Toast.makeText(
             context,
-            "Informe o código do produto!",
+            "Informe o código de um produto válido!",
             Toast.LENGTH_SHORT
         ).show()
     }
